requirements.txt
numpy
flask
gunicorn
sklearn
pandas
-e .
PyYAML
six
dill
pandas_profiling
ipywidgets
evidently
